refactor(tests): move tests/renderApp.ts to tests/utils/, rename pokemon.json to pikachu.json
