+++ react-components/src/tests/data/pikachu.json
